Crossref: use mathjax for equations
This should not matter for the generated document, as we are using the
HTML5 writer only to ensure proper escaping of special characters.
However, if this option is not set, then a warning will be created if
the math cannot be converted to a simple unicode representation. That
warning must be avoided.