Prevent special handling of `other` in citation author lists (#10)
The literal name `other` as the last author is treated as `et al.`.
However, APA style does not work well when this is used, as it always
includes the last author (preceded by an ellipses) in the bibliography
entry. We work around this by setting `others` as the family name, but
it's a hack.