Add `arxiv` target
The `preprint` target outputs a very plain file, missing any
journal-specific styling.  It's nice when the arxiv version of a paper
has similar formatting to the published version.  There have been
requests and discussions concerning a smooth way to do this.

Some issues were raised, the biggest one being possible confusion about
the publication status of a preprint.

Configure pandoc to output a LaTeX file somewhat closer to that which
would generate the official PDFs.  In order to distinguish it from a
published article, this style:

- Removes the journal logo from the header.
- Replaces the DOI and review details in the sidebar with "Preprint:
  Prepared for submission to $journal.title$".  This is borrowed from
  JCAP's LaTeX style.
- Removes the self-citation in the footer.

There are other minor differences in fonts and PDF tags because
arxiv.org only supports `pdflatex`, not `lualatex`:
- remove `\pdfvariable omitcidset=*
- remove `\setmonofont{Hack}`
- remove `\usepackage{fontsetup}`